Buttons inside of a Grid inside of Slides
Rereated the bottom-half UI functionality of the Dayton app. Now we can have icon buttons in a grid-like pattern on the botttom half of the screen. Each grid-like pattern of buttons is embedded in a Slide. This allows us to swipe left or right on it to get a new grid of buttons.